requirements.txt
bcrypt==4.3.0
pydantic==2.11.1
fastapi==0.115.12
fastapi-cli==0.0.7
python-decouple==3.8
passlib==1.7.4
PyJWT==2.10.1